split utilities file into multiple ones
