+++ app/ts/App.tsx
@@ -2,7 +2,7 @@ import { Signal, useSignal } from '@preact/signals'
 import { useEffect, useRef } from 'preact/hooks'
 import { AccountAddress } from './types/types.js'
 import { OptionalSignal, useOptionalSignal } from './utils/OptionalSignal.js'
-import { getAccounts, getChainId, requestAccounts, isAugurConstantProductMarketDeployed } from './utils/utilities.js'
+import { getAccounts, getChainId, requestAccounts } from './utils/ethereumWallet.js'
 import { DeployContract } from './ConstantProductUI/components/DeployContract.js'
 import { CreateYesNoMarket } from './CreateMarketUI/components/CreateMarket.js'
 import { ensureError } from './utils/errorHandling.js'
