+++ app/ts/CreateMarketUI/components/CreateMarket.tsx
@@ -1,5 +1,5 @@
 import { useSignal } from '@preact/signals'
-import { approveErc20Token, createYesNoMarket } from '../../utils/utilities.js'
+import { approveErc20Token, createYesNoMarket } from '../../utils/augurContractUtils.js'
 import { OptionalSignal } from '../../utils/OptionalSignal.js'
 import { AccountAddress, EthereumAddress, EthereumQuantity, NonHexBigInt } from '../../types/types.js'
 import { AUGUR_CONTRACT, DAI_TOKEN_ADDRESS, GENESIS_UNIVERSE, REPV2_TOKEN_ADDRESS } from '../../utils/constants.js'
