+++ app/ts/utils/augurContractUtils.ts
@@ -1,11 +1,9 @@
 import 'viem/window'
 import { AccountAddress, EthereumBytes32, EthereumQuantity } from '../types/types.js'
-import { createPublicClient, createWalletClient, custom, getContractAddress, http, numberToBytes, publicActions } from 'viem'
 import { mainnet } from 'viem/chains'
-import { augurConstantProductMarketContractArtifact } from '../VendoredAugurConstantProductMarket.js'
 import { AUGUR_UNIVERSE_ABI } from '../ABI/UniverseAbi.js'
 import { ERC20_ABI } from '../ABI/Erc20Abi.js'
-import { AUDIT_FUNDS_ADDRESS, AUGUR_CONTRACT, BUY_PARTICIPATION_TOKENS_CONTRACT, FILL_ORDER_CONTRACT, GENESIS_UNIVERSE, HOT_LOADING_ADDRESS, ORDERS_CONTRACT, PROXY_DEPLOYER_ADDRESS, REDEEM_STAKE_ADDRESS, REPV2_TOKEN_ADDRESS, YES_NO_OPTIONS } from './constants.js'
+import { AUDIT_FUNDS_ADDRESS, AUGUR_CONTRACT, BUY_PARTICIPATION_TOKENS_CONTRACT, FILL_ORDER_CONTRACT, GENESIS_UNIVERSE, HOT_LOADING_ADDRESS, ORDERS_CONTRACT, REDEEM_STAKE_ADDRESS, REPV2_TOKEN_ADDRESS, YES_NO_OPTIONS } from './constants.js'
 import { AUGUR_ABI } from '../ABI/AugurAbi.js'
 import { HOT_LOADING_ABI } from '../ABI/HotLoading.js'
 import { BUY_PARTICIPATION_TOKENS_ABI } from '../ABI/BuyParticipationTokensAbi.js'
@@ -16,74 +14,7 @@ import { REPORTING_PARTICIPANT_ABI } from '../ABI/ReportingParticipant.js'
 import { REPUTATION_TOKEN_ABI } from '../ABI/ReputationToken.js'
 import { REDEEM_STAKE_ABI } from '../ABI/RedeemStakeAbi.js'
 import { AUDIT_FUNDS_ABI } from '../ABI/AuditFunds.js'
-
-export const requestAccounts = async () => {
-	if (window.ethereum === undefined) throw new Error('no window.ethereum injected')
-	const reply = await window.ethereum.request({ method: 'eth_requestAccounts', params: undefined })
-	return reply[0]
-}
-
-export const getAccounts = async () => {
-	if (window.ethereum === undefined) throw new Error('no window.ethereum injected')
-	const reply = await window.ethereum.request({ method: 'eth_accounts', params: undefined })
-	return reply[0]
-}
-
-const createReadClient = (accountAddress: AccountAddress | undefined) => {
-	if (window.ethereum === undefined || accountAddress === undefined) {
-		return createPublicClient({ chain: mainnet, transport: http('https://ethereum.dark.florist', { batch: { wait: 100 } }) })
-	}
-	return createWalletClient({ chain: mainnet, transport: custom(window.ethereum) }).extend(publicActions)
-}
-
-const createWriteClient = (accountAddress: AccountAddress) => {
-	if (window.ethereum === undefined) throw new Error('no window.ethereum injected')
-	if (accountAddress === undefined) throw new Error('no accountAddress!')
-	return createWalletClient({ account: accountAddress, chain: mainnet, transport: custom(window.ethereum) }).extend(publicActions)
-}
-
-export const getChainId = async (accountAddress: AccountAddress) => {
-	return await createWriteClient(accountAddress).getChainId()
-}
-
-export function getAugurConstantProductMarketAddress() {
-	const bytecode: `0x${ string }` = `0x${ augurConstantProductMarketContractArtifact.contracts['AugurConstantProductMarket.sol'].AugurConstantProduct.evm.bytecode.object }`
-	return getContractAddress({ bytecode, from: PROXY_DEPLOYER_ADDRESS, opcode: 'CREATE2', salt: numberToBytes(0) })
-}
-
-export const isAugurConstantProductMarketDeployed = async (accountAddress: AccountAddress | undefined) => {
-	const wallet = createReadClient(accountAddress)
-	const expectedDeployedBytecode: `0x${ string }` = `0x${ augurConstantProductMarketContractArtifact.contracts['AugurConstantProductMarket.sol'].AugurConstantProduct.evm.deployedBytecode.object }`
-	const address = getAugurConstantProductMarketAddress()
-	const deployedBytecode = await wallet.getCode({ address })
-	return deployedBytecode === expectedDeployedBytecode
-}
-
-export const deployAugurConstantProductMarketTransaction = () => {
-	const bytecode: `0x${ string }` = `0x${ augurConstantProductMarketContractArtifact.contracts['AugurConstantProductMarket.sol'].AugurConstantProduct.evm.bytecode.object }`
-	return { to: PROXY_DEPLOYER_ADDRESS, data: bytecode } as const
-}
-
-export async function ensureProxyDeployerDeployed(accountAddress: AccountAddress): Promise<void> {
-	const wallet = createWriteClient(accountAddress)
-	const deployerBytecode = await wallet.getCode({ address: PROXY_DEPLOYER_ADDRESS })
-	if (deployerBytecode === '0x60003681823780368234f58015156014578182fd5b80825250506014600cf3') return
-	const ethSendHash = await wallet.sendTransaction({ to: '0x4c8d290a1b368ac4728d83a9e8321fc3af2b39b1', amount: 10000000000000000n })
-	await wallet.waitForTransactionReceipt({ hash: ethSendHash })
-	const deployHash = await wallet.sendRawTransaction({ serializedTransaction: '0xf87e8085174876e800830186a08080ad601f80600e600039806000f350fe60003681823780368234f58015156014578182fd5b80825250506014600cf31ba02222222222222222222222222222222222222222222222222222222222222222a02222222222222222222222222222222222222222222222222222222222222222' })
-	await wallet.waitForTransactionReceipt({ hash: deployHash })
-}
-
-export const deployAugurConstantProductMarketContract = async (accountAddress: AccountAddress) => {
-	const augurConstantProductMarketDeployed = await isAugurConstantProductMarketDeployed(accountAddress)
-	if (augurConstantProductMarketDeployed) throw new Error('already deployed')
-	await ensureProxyDeployerDeployed(accountAddress)
-	const client = createWriteClient(accountAddress)
-	if (!augurConstantProductMarketDeployed) {
-		const hash = await client.sendTransaction(deployAugurConstantProductMarketTransaction())
-		await client.waitForTransactionReceipt({ hash })
-	}
-}
+import { createReadClient, createWriteClient } from './ethereumWallet.js'
 
 export const createYesNoMarket = async (marketCreator: AccountAddress, endTime: bigint, feePerCashInAttoCash: bigint, affiliateValidator: AccountAddress, affiliateFeeDivisor: bigint, designatedReporterAddress: AccountAddress, extraInfo: string) => {
 	const client = createWriteClient(marketCreator)
@@ -358,6 +289,7 @@ export type ReportingHistoryElement = {
 	payoutNumerators: readonly bigint[],
 	stake: bigint
 }
+
 export const getReportingHistory = async(reader: AccountAddress, market: AccountAddress, currentRound: bigint) => {
 	const client = createReadClient(reader)
